requirements.txt
flask
flask-sqlalchemy
sqlalchemy
python-dotenv
requests
wtforms
flask_wtf